absolute minimum but it works

+++ src/main/java/com/example/demo/repository/RecordRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.repository;
 import com.example.demo.entity.Record;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecordRepository extends JpaRepository<Record, Long> {}
+public interface RecordRepository extends JpaRepository<Record, Integer> {}
